# Conflicts:
#	frontend/src/app/app.component.html

+++ frontend/src/app/modals/workgroup-edit/workgroup-edit.component.ts
@@ -1,10 +1,11 @@
-import { Component, ViewChild } from '@angular/core';
+import { Component, EventEmitter, Output, ViewChild } from '@angular/core';
 import { NgForm } from '@angular/forms';
 import { NgbModal, NgbModalRef } from '@ng-bootstrap/ng-bootstrap';
 import { ToastrService } from 'ngx-toastr';
 import { DeleteRepositoryToast } from 'src/app/deleteRepository-toast';
 import { RepositoryContainer } from 'src/app/model/RepositoryContainer';
 import { ApiService } from 'src/app/Services/api.service';
+import { RepoSwichComponent } from '../repo-swich/repo-swich.component';
 
 @Component({
   selector: 'app-workgroup-edit',
@@ -44,16 +45,25 @@ export class WorkgroupEditComponent {
   userEmail = '';
   userId = '';
 
+  repos : RepositoryContainer[];
+
    /**
     * Model Reference for closing
     */
   modalReference: NgbModalRef;
 
+  @Output() 
+  currentRepoToDeleteEvent : EventEmitter<boolean> = new EventEmitter();
+
   @ViewChild('workgroupEditModal') workgroupEditModal: WorkgroupEditComponent;
+  @ViewChild('repoSwitchModal') repoSwitchModal : RepoSwichComponent;
 
   constructor(private modalService: NgbModal, public apiService: ApiService, private toastr: ToastrService) {
     this.apiService.deleteRepositoryEvent.subscribe(() => {
       this.deleteCustomRepo();
+    });
+    this.apiService.getRepositories().subscribe(repos => {
+      this.repos = repos;
     });
   }
 
@@ -65,7 +75,7 @@ export class WorkgroupEditComponent {
     this.userId = userId;
     this.workgroupList = [];
     this.workgroupProject = project;
-    this.modalReference = this.modalService.open(this.workgroupEditModal, {ariaLabelledBy: 'modal-basic-title'});
+    this.modalReference = this.modalService.open(this.workgroupEditModal, {ariaLabelledBy: 'modal-basic-titles'});
     const header = document.getElementById('workgroupHeader') as HTMLSpanElement;
     header.textContent = 'Project: ' + project.value;
 
@@ -124,10 +134,19 @@ export class WorkgroupEditComponent {
  */
   deleteCustomRepo(){
     if(this.userEmail == this.workgroupOwner) {
+      const currentRepo = localStorage.getItem('repository');
+      if ( this.workgroupProject.value == currentRepo) {
+        this.currentRepoToDeleteEvent.emit(true);
+        this.openRepoSwitchModal();
+      }
+      else {
+        this.currentRepoToDeleteEvent.emit(false);
         this.apiService.deleteRepository(this.workgroupProject, this.userId).subscribe(res =>{
-            this.apiService.updateRepositoryEmitter();
-            this.modalReference.close();
-        })
+          this.apiService.getRepositoriesEmitter();
+          this.apiService.updateRepositoryEmitter();
+        });
+        this.modalReference.close();
+      }
     }
   }
 
@@ -142,6 +161,17 @@ export class WorkgroupEditComponent {
 
   showErrorToast () {
     this.toastr.error(this.workgroupError);
+  }
+
+  /**
+  * Opens repo switch modal
+  */
+  openRepoSwitchModal() {
+    this.repoSwitchModal.openModal();
+  }
+
+  isCurrentToDelete (value):boolean {
+    return value
   }
 
 }
